add some new features for handling errors

+++ Application/Activities/Delete.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Domain;
 using MediatR;
 using Persistence;
 
@@ -6,12 +7,12 @@ namespace Application.Activities;
 
 public class Delete
 {
-    public class Command : IRequest<Unit>
+    public class Command : IRequest<Result<Unit>>
     {
         public Guid Id { get; set; }
     }
 
-    public class Handler : IRequestHandler<Command, Unit>
+    public class Handler : IRequestHandler<Command, Result<Unit>>
     {
         private readonly DataContext _context;
         private readonly IMapper _mapper;
@@ -22,12 +23,16 @@ public class Delete
             _mapper = mapper;
         }
 
-        public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+        public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
         {
             var activity = await _context.Activities.FindAsync(request.Id);
+            if (activity == null) return Result<Unit>.Failure("Activity not found", 404);
             _context.Activities.Remove(activity);
-            await _context.SaveChangesAsync();
-            return Unit.Value;
+            var result = await _context.SaveChangesAsync(cancellationToken) > 0;
+
+            if (!result) return Result<Unit>.Failure("Failed to delete the activity", 400);
+
+            return Result<Unit>.Success(Unit.Value);
         }
     }
 }
